Move IsLocalAddress outside of the config package
This commit also make a module out of pkg/util/system.
