+++ pkg/config/setup/ipc_address.go
@@ -7,35 +7,11 @@ package setup
 
 import (
 	"fmt"
-	"net"
 
 	pkgconfigmodel "github.com/DataDog/datadog-agent/pkg/config/model"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"github.com/DataDog/datadog-agent/pkg/util/system"
 )
-
-// IsLocalAddress returns the given address if it is local or an error if it is not
-func IsLocalAddress(address string) (string, error) {
-	if address == "localhost" {
-		return address, nil
-	}
-	ip := net.ParseIP(address)
-	if ip == nil {
-		return "", fmt.Errorf("address was set to an invalid IP address: %s", address)
-	}
-	for _, cidr := range []string{
-		"127.0.0.0/8", // IPv4 loopback
-		"::1/128",     // IPv6 loopback
-	} {
-		_, block, err := net.ParseCIDR(cidr)
-		if err != nil {
-			return "", err
-		}
-		if block.Contains(ip) {
-			return address, nil
-		}
-	}
-	return "", fmt.Errorf("address was set to a non-loopback IP address: %s", address)
-}
 
 // GetIPCAddress returns the IPC address or an error if the address is not local
 func GetIPCAddress(config pkgconfigmodel.Reader) (string, error) {
@@ -58,7 +34,7 @@ func getIPCAddress(cfg pkgconfigmodel.Reader) (string, error) {
 		key = "cmd_host"
 	}
 
-	address, err := IsLocalAddress(cfg.GetString(key))
+	address, err := system.IsLocalAddress(cfg.GetString(key))
 	if err != nil {
 		return "", fmt.Errorf("%s: %s", key, err)
 	}
